+++ src/main.c
@@ -2,7 +2,7 @@
 #include <stdlib.h> 
 #include <string.h> 
 #include <sys/types.h>
-#include <sys/stat.h>
+#include <sys/stat.h>	
 #include <unistd.h>
 #include <sys/mman.h>
 #include <fcntl.h>
