+++ src/parseur/headerExpect.c
@@ -1,0 +1,68 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include "../tree.h"
+#include "subroutineMultipleTerminal.h"
+#include "headerExpect.h"
+
+
+#define true 1
+#define false 0
+
+int Expect_header(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = case_insensitive_string(now, req, fils = createFils(pere), "Expect"))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = case_insensitive_char(now, req, fils = createFrere(fils), ':'))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
+        purgeFilsAndFrere(fils);
+    }
+    else {
+        fils = createFrere(fils);
+    }
+    now += len;
+
+    if(!(len = Expect(now, req, fils))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    if(!(len = OWS(now, req, fils = createFrere(fils)))) {
+        purgeNode(fils);
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "Expect_header", pere);
+    return now-p;
+}
+
+int Expect(int p, const char *req, node *pere)
+{
+    int len;
+    int now = p;
+
+    node* fils;
+
+    if(!(len = case_insensitive_string(now, req, fils = createFils(pere), "100-continue"))) {
+        purgeNode(fils);
+        return false;
+    }
+    now += len;
+
+    putValueInNode(p, now-p, "Expect", pere);
+    return now-p;
+}
