+++ src/parseur/headerExpect.h
@@ -1,0 +1,10 @@
+#ifndef __HEADEREXPECT_H__
+#define __HEADEREXPECT_H__
+
+/*Prototype*/
+
+int Expect_header(int p, const char *req, node *pere);
+
+    int Expect(int p, const char *req, node *pere);
+
+#endif
